publisher-service/requirements.txt
fastapi
pika
python-dotenv
uvicorn
datetime
python-multipart
